requirements.txt
django
django-cors-headers
weasyprint
pillow
djangorestframework
gunicorn
pdfkit
pdfkit

fpdf
